requirements.txt
langchain-community
langchain_openai
sentence-transformers
faiss-cpu
langchain-huggingface
openai